Add: ✨ Defining table for the column to prepare to showcase the UI
One important thing: Even though we have data contains 8 column from the mock data.
The email and age is not included. This is to simply to showcase that you can pick what data needs to be rendered in the UI
If a column left out from this column array, it won't make it into the UI.